app/requirements.txt
Django
djangorestframework
django-admin-list-filter-dropdown==1.0.3
django-cors-headers

Pillow==9.3.0

django-storages
boto3

psycopg2-binary==2.9.5
django-dbbackup

drf-spectacular
unidecode
rollbar
gunicorn

redis
celery[redis]
django-celery-beat
django-redis
flower

django-debug-toolbar
